Fix linting issues by moving mocks to a single file.
